requirements.txt
marker-pdf
transformers==4.45.2
pillow
markdown